api/requirements.txt
fastapi
uvicorn
psycopg2
python-multipart
Pillow